mutants/requirements.txt
Django
djangorestframework
django-cors-headers 
PyMySQL==1.0.2